Use badfilecontexttitle to set targetlang for imageinfo requests
Parser::makeImage always sets targetlang based on the page language.  In
order for Parsoid to support that in standalone mode, we set it based on
the overloaded badfilecontexttitle, which is the only context we have
when making imageinfo requests.

Follow up to I348840ef405e1370cc0c17d69051bce30153c9c0

Bug: T208620
Change-Id: Id01b84907c6817194c53dbb9d731e5383e7a0056