Improve handling of class-level test setup failures in run()
Signal that the test run needs to stop and record the exception via
addError() so that it will be displayed in the output and the exit
code will reflect test failure.

Previously, errors would cause the tests to suddenly stop, and any
buffered output about previous errors, warnings, and failed tests
was lost.

Bug: T292239
Change-Id: I5b2a7bcbb38760fcd10fccfbf000f4a8511f8774