SuiteEventsTrait: don't call setUp() for an empty suite
SuiteEventsTrait is a hack to support the parser test suite classes. If
the suite is empty due to e.g. group filters, doing setUp() and
tearDown() is unnecessary and expensive (requiring DB cloning). Also,
errors from setUp() or tearDown() on an empty suite were not properly
handled.

So if the suite is empty, do the same thing as TestSuite::run(), leave
the result unmodified.

Bug: T292239
Change-Id: I0828257d6dd0bbc5b1633afde5ff162e96169675