Span representing broken media element is only expected as first child
Video.js appends a bunch of spans in the figure which this selector is
being applied to, so we attempt to limit its scope.

A follow up introducing a class for this is forthcoming, but this should
improve things in the meantime.  The class won't be present on all the
cached Parsoid content even after we roll it out there so this will
continue to be useful until that's all regenerated.

Bug: T304010
Change-Id: Ic2699ab1d1d0160601d2beb48c34b556d286e84c