Do not hide the "time" element if it also carries the "date" role
If changes are grouped by date, the lines display only the time, and this time is also assigned the general-purpose "date" class.

Therefore, we should hide the time only when the changes are not grouped by date, and there is a separate date element (which also contains the time) next to the time element.

Bug: T304406
Change-Id: Ib9f854b193babce51136899bc11ca12231475ffc