SettingsBuilder: add comments about batched caching.
One reason to have batching in SettingsBuilder is so we can cache an
entire batch of settings after merging them, rather than caching each
file individually. However, this has not been implemented, and it's not
entirely clear if it's a good idea. This patch adds comments to
SettingsBuilder that point to the relevant ticket, for reference
and consideration when working on SettingsBuilder in the future.

Bug: T304493
Change-Id: I1fb7b7e4a97e6d0788c9b1732a1718f3c8b365d5