RollbackPage: Make rollback not overwrite manual RC patrol status
When performing a rollback, we marks the reverted edits as
patrolled (on UseRCPatrol wikis) and also pretend they were by a bot
(discussion at T304766).

The way it is marked as patrolled is by setting the patrol status to
autopatrolled. But, if (one of) the edits was already patrolled before
the rollback, the manual patrol status should not be overwritten.

* Fix this.
* Also improve query performance by setting via rc_id instead of
  through more complex conditional clauses in the update query
  (should help with T260798).
* Test as much of this patrolling stuff as possible.

Bug: T304766
Bug: T260798
Change-Id: I66bb6575dad4a970339e93b91fe486727111092a