Handle webm files with AV1 and/or nullbyte terminated VP8/9
There is a bug in several muxers where it writes a null terminated
string into the field for the codecname. This increases the length of
the field by one, and thus we get the sequence 0x86 0x86

It is not valid to have null terminated strings, but as ffmpeg and
everything else seems to deal with these files, we probably should as
well.

Bug: T226311
Change-Id: I6f9796018a5a31f02d9d2707a3e31cf57674deb6