Update Media and File namespace names for Swahili
The "File" namespace was defined as "Picha", which means "Image",
and the "File talk" namespace was defined as "Majadiliano_ya_faili",
which actually means "File talk".

The "Media" namespace was defined as "Faili", which means "File".
The definitions of "Media" and "File" are obviously a mistake.

"Faili:" was used on eight pages in the Swahili Wikipedia
and on one page in the Swahili Wiktionary. In all of them
it was used incorrectly, and I corrected them all. Now it doesn't
appear on any pages in any Wikimedia project.

The "Media" namespace can have the same name as in English,
so that line is removed here. File now actually becomes "Faili".

Bug: T304920
Change-Id: I742798459a6b3f2141bccf180d342f1c2055a932