phpunit: Remove Makefile
To the best of my knowledge, nobody uses this, and it is yet another
partial summary with some commands to run in some cases.

We have "composer" run commands for most of these, and documentation
pages and `phpunit --help` for the rest. Given lack of use, I have not
copied the remaining unique commandds here to composer.json, but
patches for that would be welcome.

Change-Id: Idde27dfe266295e6c8e4aeefaaf9335c2014e1b2