phpunit: parse any --boostrap parameter in getopt() call
This is needed in case such a parameter exists so that --configuration
will be recognized. This is due to how getopt() stops looking once
it finds an unrecognized field.

So long as the order is "configuration" first, the other options all
work and get passed through to phpunit. But, paratest specifies
bootstrap before configuration, which means this subclass won't see
"configuration".

Follows-up 204776e42c546b, after which stopped working.

Bug: T50217
Change-Id: I47fa647a9b26f099f6a399f256d8154bb20c4c17