resourceloader: Disable mw.loader.store on safemode pages
On special pages where disallowUserJs() is used, or when setting
safemode=1 by hand to index.php, disable mw.loader.store.

OutputPage already normalizes disallowUserJs() to safemode=1 as
passed on to ResourceLoaderClientHtml and ResourceLoaderContext.

Use it in the startup mode to disable this feature.

Invert the internal substitution variable to avoid confusing
double negatives, and to match the config name.

Bug: T145498
Change-Id: I2fdba968b4703ec8826bdaf91340bdead058f02b