Make old vs. latest revision more obvious in ParserOutputAccess
The same check is done in multiple places, just slightly different
every time, but always boiling down to the same idea: The latest
revision is requested either when it's specifically requested or
missing. Otherwise it's an old one.

The only purpose of this patch is to make the code easier to read.
I made sure there is zero functional change.

Bug: T304813
Change-Id: I0a5c84eb137dfbbefb1ef57eaf8711971b991911