Untangle dependencies between PoolWorkArticleView subclasses
The base class should not need to know anything about caching.

The motivation for this patch is to loosen the strong dependencies
between these subclasses, to possibly turn them into proxies or
something with much looser coupling.

This patch doesn't change any behavior. The code is just moved to a
slightly different place, but executed in the same order.

Bug: T304813
Change-Id: Icd68538c85c193c3d17443154bfdf6d5bce7661c