Remove some dynamic config from Setup
The default value for GitInfoCacheDirectory doesn't seem to be necessary
outside of GitInfo.php, so move it from Setup to there.  Theoretically
this could break third-party code that uses this setting and was relying
on Setup.php to set a default value, but it doesn't seem very likely.

I didn't find any other dynamic configuration that could be moved out of
Setup so easily:

* MainWANCache isn't accessed outside of ServiceWiring, but Setup
  still needs to check it to initialize WANObjectCaches.
* ForeignFileRepos is only read in FileBackendGroup, but various tests
  (including in extensions) reset it and their behavior will be affected
  if they no longer override the various related settings like
  UseInstantCommons.
* ShortPagesNamespaceBlacklist handling could be moved to
  SpecialShortPages.php, but that would require re-adding it as a
  recognized configuration setting, which doesn't seem worth the tiny
  code simplification.

Bug: T305093
Change-Id: Iec8c391fec8396846216a9708d75df57f7d799a2