WebRequest: Micro-optimise getIP
Called on every request. Profiling attributes this to the no-op
for the wfDebug() call in Setup.php, but in practice there are later
calls as well.

Use str_contains() while at it. This isn't where the time is spent
and isn't native until PHP 8, but might as well start using it.

Bug: T85805
Change-Id: Ia7117e6d71327edd396b0cc0e59d90af158d381e