mediawiki.special.edittags: Move into misc-authed bundle
This is a small script, needed for a relatively minor logged-in
feature. It has no perf budget and is not highly-visible like edits,
pageview, RC/WL, or signup. Bundle it with misc instead of its own
separate entrypoint.

This is mostly to reflect the feature's other aspects which are
also integrated into broader buckets and not major codebases that
stand on their own. It is motivated by other bundles recently
created in core and in other WMF extensions.

Test plan:
* Have or create at least one user-defined tag via Special:Tags.
* Browse any action=history page while logged-in as admin.
* Tick a box and click "Edit tags".
* It should continue to enhance the native multi-line HTML select
  to a single-line input with tag cloud and autocomplete.

Change-Id: I88806cf80fa4a3230ae9239b710f413e11fcc869