changetags: Remove "hits" key from exported RCFilters module data
Fix the `mediawiki.rcfilters.filters.ui` version churn.

While at it, simplify the code to no be structured like the hit count
plays a central value, and instead iterate the list directly and
check the value in the loop.

Also, we can reuse listDefinedTags() for consistency with other
code in this class.

Bug: T298520
Change-Id: Ib6e1369aae6e616be66408809f34ee50b441ef4b