Sort logging table by timestamp and log id
Sorting on log_id must not give a result in time order,
because maintenance scripts can insert newer ids with older timestamps.
When only sorting for log_timestamp the result is not stable,
because for items with same timestamp the order is not defined.
For Special:Log the same timestamp gives problems on pagination

Bug: T305590
Change-Id: Ic0c2ebd19ee6a8aa570c5a12a92b31008135491c