docs: Use the default EXTRACT_PRIVATE=NO for Doxygen output
This will remove some noise from long classes where currently these
members are listed before the public methods, which makes the docs
harder to navigate. The audience of doc.wikimedia.org where we
show this, I think is not likely to be interested in private members
and methods. That can always be found when navigating within a text
editor of course, where one is likely to interact with private methods.

For example, at https://doc.wikimedia.org/mediawiki-core/master/php/classWANObjectCache.html
the private members are before the main class docs render.

Change-Id: I8fda01f653e9cec311ecb10df98c3f871de267e6