profiler: Deprecate unused 'trace' option
This has no usage in core or in Codesearch Everywhere.

The creation of logical profile entries (such as for templates in
the parser, and for DB queries) remains supported and this is how
the class is used in core. Capturing detailed traces is only supported
via implementations like ProfilerXhprof and ProfilerExcimer.

Bug: T305100
Change-Id: Ib06ec25bfa1f9565ce85b54e8375c0a5f99a912a