Make PageUpdater accept PageIdentity
Internally it still uses WikiPage. While this change shouldn't have any
effect in theory, this is not guaranteed because WikiPage is not
stateless and if different parts of the code use different instances we
may see unexpected errors. Hence, this patch is limited to the bare
minimum so that it's easier to revert if necessary.

Change-Id: I2c18a62ccc04720a2ef25298515850e91fb288e9