cache: Introduce member to expose service options for configs
There are some static methods in sub-classes of FileCacheBase
that use the service locator to get the main config object
which maybe I think we can just use DI in the methods to
pass this from the calling code.

That could be in a follow-up patch, for now, we can leave
things as they are here?

Change-Id: Ic75e1e5f867faee7fc71c34552f9f0a7a124d7ae