Fix partly broken showMissing option in TitleInputWidget
There is a "showMissing" option that's enabled by default. What it does
is to always show the user's input as the first search result, even if
a page with this name doesn't exist. It's shown in red then. This allows
the user to simply press enter to accept the current input, no matter if
a page with this name exists or not.

This option ignores the namespace option. It always searches in the main
namespace. (Except the user pasted a prefixed page name including the
namespace.) Results from other namespaces are dropped further down in
the same widget, in .getOptionsFromData(). The consequence is an extra
API request that turns out to be useless in almost all cases.

1. Make the showMissing option respect the namespace.

This causes another problem: Now, every UI that uses this or a subclass
of this widget together with the namespace option will potentially start
offering this extra non-existing result.

2. Disable showMissing by default when the namespace option is used.

This roughly restores the previous behavior.

Alternatives to this proposal:
* Fix it and wait for incoming reports about unexpected red results.
* Always disable showMissing by default.

Change-Id: I61d09d3b969c1d19d113f29f93c34bb028cedc7a