Use the same tooltip for transcluded sections as normal ones
Remove the code that outputs self-closing <mw:editsection ... /> tags
in Parser, previously used for transcluded sections.

Remove the ability to handle them in ParserOutput. We don't need
backwards-compatibility with cached content, because that feature did
not work correctly for several years: Remex-Tidy always expanded them
to normal open and close tags.

Remove handling for this case in skin code (and fix documentation).
These are backwards-compatible changes.

Depends-On: Idbf0b95a3c0b04caa056b71dd08f46659920114a
Bug: T306299
Change-Id: I3fac0f34d134d8eec46c7eefa3ad2b67abb957da