mediawiki.skinning: Make Form UI legend font consistent
The font size for UI Form legends are not consistent on all pages
because HTMLForm overrides this CSS rule in line 16 of
resources\src\mediawiki.htmlform.ooui.styles.less.
Therefore the legend size is different between pages
Special:Move and Special:Preferences. This removes the rule so it
should default of 100% all of the time.

Change-Id: I34656ca7aebf507c031e28b089f1a8254a6dfad8