api: Add support for LinksMigration in ApiQueryBacklinksprop
It should be generalized enough to be used in future tables as well.
Note that it removes the useIndex when the table is in mapping even if
it's read old because useIndex is evil and should not be used like this.

Also removing the straight join hint as it would intefere with
linktarget join and I tested the query with the current version of
mariadb we have in production and it works just fine.

Bug: T304780
Change-Id: Iea3e0b309ddd013687008571513afeef6b9b672b