Fix broken user name vs. IP import logic in WikiImporter
1. The code that processes $logInfo first calls setUserIP(), and then
uncontitionally calls setUsername() with "Unknown user". Because
both setters write to the same field (see I6687305) the IP was never
stored.

2. Flip conditionals to prefer user name over IP. Usually this
shouldn't make a difference. Only one of the two fields should be
set. But if both are set the better logic is to store the name of the
logged in user, instead of silently turning them into an anonymous
one.

I am not sure if this is actually related to the linked tasks,
but might be. The bug was introduced in 2015 via Idf95263 (T121338).
The setUserIP/setUsername problem exists since 2004, see
https://phabricator.wikimedia.org/rMW436a0280.

Bug: T11041
Bug: T236644
Change-Id: I1b8ed89a4190b55d3a9e4ecd9a83b0a042e493be