Clean up tables of Chinese conversion rules
Most of the changes are generated with the help of my modified version
of Makefile.py and the BCompare software.
Also:
* Rules about 鰛、鳁、鰮 in symme_supp.manual are removed since they are
not one-to-one rules as the file should handle.
* Rules of 乾卦、髮踴沖冠、矇松雨 in tradphrases.manual are removed to fix
the output table.
  To reviewer: You may want to check other terms which are removed from
tradphrases_exclude.manual but still exist in tradphrases.manual.

Since some useless rules are added in recent years, I will upload my
revised version of Makefile.py in a follow-up which can provide warnings
about probably wrong inputs.

Change-Id: I139e26f7b6fc5730ae65474567f5a53c77e3d292