Setup.php: allow loading config schema from MainConfigSchema.
Loading the config schema from MainConfigSchema using reflection is
slow, so we don't want to do it when serving traffic. But when
re-generating DefaultSettings.php or config-schema.php, we want to load
the schema from the canonical source. That way, these scripts can still
be used if DefaultSettings.php and config-schema.php are broken.

Change-Id: Idc0d8ca0f4a700b771e6182d53ef4a0efab7110c