Remove hiding of extra namespace checkboxes on legacy watchlist and RC
In the legacy version of Special:Watchlist and Special:RecentChanges
(available by enabling the preferences named "Use non-JavaScript
interface" on their respective tabs), there's a dropdown "Namespace"
and two checkboxes "Invert selection" and "Associated namespace",
which are hidden when the default value of "all" namespaces is
selected.

It doesn't seem worth it to maintain this feature. The checkboxes are
now always shown, even when "all" is selected.

The code implementing this was always loaded, thus causing a small
performance hit for all users, even when the new interface is enabled
(and even on other pages, since it was a separate ResourceLoader
module).

Change-Id: Ia5c4790358c5e631c2305343c8bdfddae15ce12b