UserAuthority: Fix wikitext escaping for block errors
Status and Message behave differently when given a list of message
parameters. Since the errors generated by PermissionManager is intended
to be used by Message, we need to update UserAuthority to do that.

Fixed StatusValue::replaceMessage to support MessageSpecifier objects
like StatusValue::hasMessage does. This is so that the block message
formatting in ApiBase does not break.

Bug: T306494
Change-Id: I26405c680839b4679709e750d6d54e699c1efa66