Add HtmlHelper::modifyElements() for small HTML modifications
Adds HtmlHelper, which is intended to be a static utility class
for doing simple modifications to HTML strings in a safe way,
without exposing the caller to the complex Remex interface.
Currently only has one method, which is mainly useful for
modifying element attributes.

The code is largely based on the Wikibase FormatEntities class.

Bug: T217850
Change-Id: I45db9e61e47eb69df32a167d9d1dd146a8719676