Fix help message for "Recent changes → Use non-JavaScript interface" preference
This preference only applies to Special:RecentChanges and
Special:RecentChangesLinked, but not to Special:Watchlist,
which has its own preference on the "Watchlist" tab.

It could theoretically apply to similar pages (anything that extends
ChangesListSpecialPage), but there aren't any, so remove that part.

Add links to the relevant special pages in the messages, because I
think that Special:RecentChangesLinked may not be known to most users.

Change-Id: If6b402938247bd76ccb0b99c62cc60ad79efc3e7