SpecialWhatLinksHere: Convert form to OOUI
The current form is changed to OOUI with some CSS rules introduced.
The namespace select and invert checkbox are grouped in a section
with a flexbox to keep them in the same line if possible. Filter
checkboxes are set to be inline-block for compact spacing. The whole
form is also collapsible and will be collapsed when user is browsing
results (i.e. 'target' parameter is set).

Bug: T117754
Change-Id: Iebdec0dd3618fe4ee91514c082c6e63640cd8b01